refactor(film-detail): split monolithic 1587-line file into focused modules
Extract modal dialogs from FilmDetailScreen.tsx (1587 → 477 lines) into:
- film-detail/types.ts: shared types (ActionType, ActionData, EditData, ModalBaseProps)
- film-detail/TransitionModals.tsx: load, finish, partial, reload modals
- film-detail/DevScanModals.tsx: send-to-dev, develop, scan modals
- film-detail/EditModal.tsx: full film edit modal

The main FilmDetailScreen now focuses on layout and state orchestration
while each modal group is independently maintainable.
